requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Cache
redis==5.0.1
hiredis==2.2.3

# Exchange integration
ccxt==4.1.92

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Technical analysis
TA-Lib==0.4.28

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
torch==2.1.2
optuna==3.5.0
lightgbm==4.1.0

# Quantitative finance
zipline-reloaded==3.0.4
quantlib==1.32
pyfolio==0.9.2

# Time series
statsmodels==0.14.1
arch==6.2.0

# Async/concurrency
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration
python-dotenv==1.0.0
dynaconf==3.2.4

# Utilities
requests==2.31.0
schedule==1.2.0
click==8.1.7
rich==13.7.0
python-multipart==0.0.6
python-dateutil==2.8.2
orjson==3.9.10
arrow==1.3.0
psutil==5.9.6
cerberus==1.3.4
joblib==1.3.2
threadpoolctl==3.2.0
httpx==0.25.2
asyncio-throttle==1.0.2
aioredis==2.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# API documentation
openapi-generator-cli==7.0.1

# Deployment
gunicorn==21.2.0
docker==6.1.3

# Security
cryptography==41.0.8
python-jose==3.3.0
passlib==1.7.4
